+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # NEXT RELEASE
+* **(BREAKING CHANGE)** Export TheAlgorithm as the default export from fedialgo instead of a named export
+* Add displayNameFullHTML() to Account object
+* (Demo App) Add icon link to show the raw Toot JSON in a modal
 
 ### v0.37.2
 * Add `getCurrentState()` method to `TheAlgorithm`